refactor: shuffle classes around
